Styles enhancements.
Include bootstrap-compass, which includes some helper functions for
Compass. Include a commented out copy of bootstrap's variables to allow
for modification of the look and feel of bootstrap.

Gulp was running Sass files that began with an _ through Compass. That
shouldn't be the case. _ files in Sass are only for importing to other
Sass files.