Bower
Use Bower to manage CSS/JS pacakges. Removed Modernizr file and now
managing it with Bower.
Add jQuery with Bower.

The scripts:vendor and styles:vendor gulp tasks will copy Bower files
for use in building and serving. It will save them to a vendor/
subdirectory. See index.html for an example of Modernizr and jQuery.